move deps
