+++ extensions/amp-date-countdown/1.0/component.js
@@ -3,7 +3,8 @@ import {dict} from '#core/types/object';
 
 import * as Preact from '#preact';
 import {useEffect, useMemo, useRef, useState} from '#preact';
-import {Wrapper, useRenderer} from '#preact/component';
+import {Wrapper} from '#preact/component';
+import {useRenderer} from '#preact/component/renderer';
 import {useAmpContext} from '#preact/context';
 import {useResourcesNotify} from '#preact/utils';
