+++ extensions/amp-timeago/1.0/component.js
@@ -4,7 +4,8 @@ import {getWin} from '#core/window';
 
 import * as Preact from '#preact';
 import {useCallback, useRef, useState} from '#preact';
-import {Wrapper, useIntersectionObserver} from '#preact/component';
+import {Wrapper} from '#preact/component';
+import {useIntersectionObserver} from '#preact/component/intersection-observer';
 import {useMergeRefs, useResourcesNotify} from '#preact/utils';
 
 import {format, getLocale} from './locales';
