+++ src/preact/component/index.js
@@ -1,6 +1,7 @@
+// ⚠️ Be mindful of what's exported from this file.
+// All exported symbols end up in the shared bento.js file.
+
 export {ContainWrapper} from './contain';
 export {Wrapper} from './wrapper';
-export {useRenderer} from './renderer';
 export {useValueRef} from './value-ref';
 export {useDOMHandle} from './dom-handle';
-export {useIntersectionObserver} from './intersection-observer';
